# docs/swagger.yaml
definitions:
...
  main.errorResponse:
    properties:
      error:
        type: string
    type: object
...
  store.User:
    properties:
      activated:
        type: boolean
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      roles:
        items:
          type: string
        type: array
      username:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for a social media application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Social API
  version: "1.0"
paths:
...
  /users/{userId}:
    get:
      consumes:
      - application/json
      description: Get a user by their unique ID
      parameters:
      - description: User ID
        format: int64
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: Retrieve a user by ID
      tags:
      - users
  /users/{userId}/follow:
    post:
      consumes:
      - application/json
      description: Follow a user by their ID
      parameters:
      - description: User ID to follow
        format: int64
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/main.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.errorResponse'
      security:
      - BearerAuth: []
      summary: Follow a user
      tags:
      - users
  /users/{userId}/unfollow:
    post:
      consumes:
      - application/json
      description: Unfollow a user by their ID
      parameters:
      - description: User ID to unfollow
        format: int64
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.errorResponse'
      security:
      - BearerAuth: []
      summary: Unfollow a user
      tags:
      - users
  /users/activate:
    post:
      consumes:
      - application/json
      description: Activate a user account using the provided token
      parameters:
      - description: Activation token
        in: body
        name: activation
        required: true
        schema:
          properties:
            token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: Activate a user account
      tags:
      - users
...
  /users/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return a JWT
      parameters:
      - description: Credentials
        in: body
        name: credentials
        required: true
        schema:
          properties:
            email:
              type: string
            password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: Login
      tags:
      - users
  /users/register:
    post:
      consumes:
      - application/json
      description: Register a new user with username, email, and password
      parameters:
      - description: User registration data
        in: body
        name: user
        required: true
        schema:
          properties:
            email:
              type: string
            password:
              type: string
            username:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: Register a new user
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: 'Provide the JWT token as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"